Cutting down database

+++ docs/sphinx/source/clean.py
@@ -11,12 +11,13 @@ here = Path(os.path.abspath(__file__)).parent  # docs/sphinx/source/
 scope = Path('C:/Users/rober/Documents/Research/RomCom/bin/docs')
 
 
-def empty(path: Path, preserve: List[str]):
+def empty(path: Path, preserve: List[str], confirm: bool = True) -> Path:
     """ Remove all files and directories contained in ``path`` except those listed in ``preserve``.
 
     Args:
         path: The folder to empty
         preserve: The files to preserve in `path`.
+        confirm: ``True`` to require confirmation of every deletion.
 
     Returns: ``path`` which now contains only ``preserve``.
     """
@@ -24,12 +25,12 @@ def empty(path: Path, preserve: List[str]):
         if filename.name not in preserve:
             if filename in here.parents:
                 raise UserError(f'Cannot delete {filename} as it contains the clean.py script.')
-            if not input(f'Press return to delete {filename}'):
+            if confirm and not input(f'Press return to delete {filename}'):
                 Store.delete(filename)
             else:
                 raise UserError('User aborted cleanup.')
 
 if __name__ == "__main__":
-    empty(scope / 'sphinx' / 'source' / 'pages' / 'api', preserve=['api.rst'])
+    empty(scope / 'sphinx' / 'source' / 'pages' / 'api', preserve=['api.rst'], confirm=False)
     empty(scope / 'sphinx', preserve=['source'])
     empty(scope, preserve=['sphinx', '.nojekyll'])
